requirements.txt
python-telegram-bot==12.7
boto==2.49.0
requests==2.25.1